apps/ai-engine/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
psycopg2-binary
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
